Bug #11865: possible to download duplicate YouTube videos
The add_download() function was running the flash scraper to get the actual URL
to download.  However, this confused the downloader, since each run through the
scraper returns a different URL for the same video.  Since the downloader also
runs the flash scraper, we don't have to do it in the add_download() call.
This gives the Item the correct URL in the DB, but still downloads the file
correctly.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9734 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 4fa9c99792aafc5972b8a19584c0afb59961e883