Bug #11879: Miro guide isn't displaying
gtkmozembed was not being correctly set up, due to two similarly named
functions to set the path to the Mozilla libraries.  We switched the order that
the methods are checked, but this caused errors w/ python-gtkmozembed 2.25.  By
defaulting to set_comp_path, we work on both 2.19 and 2.25.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9735 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: f5de737ded96caed29b92361c81321d2f2bee127