Bug #11907: autodownload setting doesn't change immediatly
The items were not being notified by the database that their autodownload
status had changed, because the feed wasn't joined to the Item view and change
messages weren't sent for related items anyway.

Unfortunately, this sends a 'changed' message for every item in the view; I
couldn't think of a way to actually check if the item was modified.  Only
updating the auto download state seems to hit this code path though, so it
shouldn't be too much of a performance hit.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9736 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 7f0a33b448d6320f5eca1d156e235388babed418