Optimization for bulk channel folder adding.
Try to avoid calling signal_change on all the items that are being updated.
This is slow for many reasons, but the main one is that we have to instantiate
all of the items from disk at once.  Instead, only update the views that care
about folders -- as far as I can tell, this is always 0 at this point.

Made the API to get ViewTrackers for a DDBObject class public.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9743 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: aa40d7490235224233c671e61ea8dedfc3981c66