Fix for #11936
This was a combination of the database tracker stuff working "wrong" and the
test being old.  I use the scare quotes, because it's not exactly clear how
the database stuff should work.

Now the way it works is that when we call signal_related_change(), every item
in the view that isn't added or removed, gets the 'changed' signal.  Updated
the test to take advantage of that.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9812 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: c1830a1c9c5a96108ad13866dbe82583609ce372