Fix for #11941
Moved code to check for when all a container item's children have been deleted
from remove() to expire().  This fixes the problem, which was when the
container was removed, it would remove all it's children, which would cause it
to be removed a second time.

expire() seems like more of a high-level function, so it makes sense to me
to move the logic there.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9819 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: dfc0763bd69391fb41b22efed5e06c385c9f2279