Bug #11949.  This redraws all the items in a playlist when Miro starts playing an item, switches from one item to the next, or stops playing an item.
It'd be nice if instead of redrawing all the items we could instead
redraw the item being played and the previous item, but the playback
module is pretty complex and I didn't want to refactor things in
order to determine the previously playing item.  We can do that some
other day.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9827 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 9f29f4fcdd90c050ed29dd3296bff676c4980ac7