Fix for #11949
The issue was that we were using the backend code to handle something that's
only known to the fronted.  The backend has code to optimize items changing --
if the ItemInfo object doesn't change, then we don't send a change message to
the fronted.  Since the frontend is the component that handles the currently
playing item, the backend has no way to know about that and ends up not
sending the change messages.

Instead, have the item list controllers check whenever the currently playing
item changes.  If so, redraw the entire item view.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9834 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 09018bd45775c0ab8a1c3713590a157cbe0ea974