Fix for #11971.
The issue was problems running updates through SQLite when the disk was full.
This would initially cause "disk is full" errors and also cause "sql logic
error or missing database" errors when we tried to commit.

I added code that worked similar to the 2.0 code for handling these things.
Basically, when we see an OperationalError when trying to perform an update,
then we tell the user we think their disk is full and save the SQL statement
for later.  Then we periodically re-try the SQL.  If at some point in the
future, we can successfully execute it, then we tell the user we think the
disk isn't full anymore.

One snag was that I couldn't get the "sql logic error or missing database"
exceptions to go away when we tried to commit, even if the disk was no longer
full, we called rollback(), etc.  The way around it was to go back to
auto-commit mode like in the last version (the isolation_level=None parameter
to sqlite3.connect).

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9861 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 2aa0c0036cb91004da7909ea6ae8a765129bc38e