Possible fix for #12001.
In upgrade89, it's possible that we set seen and keep to NULL, which is a bad
value.  They should be set to False, which we store in the database as 0.

Not sure if this will fix the problem, but I think it will.  It's a good
change in any case.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9896 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 58d71f2f7116d575920402766d3531ced749ae2d