Merging r9895 and r9896 into Miro-2.5 branch.
r9895

Added a progress dialog for DB upgrades.

Made a new module called dbupgradeprogress, which other modules use
periodically while they're doing the db upgrade work.

Made it so that eventloop.startup() quits before the handlers for
'thread-will-start', 'thread-started' and 'thread-did-start' are invoked.
This lets us create a dialog box before the DB upgrades are finished running.

This code was started by will and finished by me. 

r9896

Possible fix for #12001.

In upgrade89, it's possible that we set seen and keep to NULL, which is a bad
value. They should be set to False, which we store in the database as 0.

Not sure if this will fix the problem, but I think it will. It's a good
change in any case. 


git-svn-id: https://svn.participatoryculture.org/svn/dtv/branches/Miro-2.5@9897 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 9e4dc9b1af430f004661e36c7b16f1574e46f685