Bug #11884: race condition between deleted items messages
It appears that deleting a watched item or a site can be involved in a race
condition which causes messages to arrive in the wrong order.  This patch
catches the error raises when we can't find the object we're trying to
hide/remove.

git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@9920 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: fd762b943f481e7e33a4f1271eb24d6f2e117356