Bug #12150.  Fixes a corner get_filename problem.
Previously, get_filename would return FilenameType('') if
self.filename is None.  But that's a total bogus filename and there's
no way to programmatically determine whether it's a valid filename
or not.  I removed the check in the code so now if self.filename is
None, it returns None.
In doing this, I had to add a check to split_item so it doesn't
just plow through things without checking to see if it's a valid
filename or not.
It's possible this will uncover other issues where the code was
blindly assuming it had a valid filename.

Original sha1: b3af1d486b96733041281673a3f3a1d451b31c1b