Added code to better deal with OperationalErrors (#12101)
Made it so when we see them we stop doing things and wait until they go away,
instead of trying to procede in spite of them.  This should prevent errors
like #12108.

Added code to allow the backend to block waiting for a dialog to return.
Added code for the backend to tell the frontend to quit.

Original sha1: b18d50fed3be0bb083b918b5bc44b67cbbc3a3c3