Reworked video box for gtk platforms.
I rearranged items in the video box for gtk platforms so that the
Expired text and Keep link are to the left of everything rather than
on top.  This fixes the problem caused by having an expiration of
never where the Delete and Share | Permalink links are aligned to
the bottom with a big space above them.
I'm pretty sure there was a bug for this, but I can't find it.

Original sha1: a21c8a68ec3d88f38287a3e2bd17ea4394551f2d