Bug #12184.  Validate tab_ids data before using.
If the tab_ids field is parseable python, but evals to something that
isn't a list of ints, then it causes problems down the road.  This adds
type-checking code to setup_restored.  If the values aren't of the
correct type, then we change it to [].

Original sha1: de65d859fc6844e0bdc5f9858c0a25762dbed0d1