Bug #12185.  Set up the tab_ids list correctly.
If the tab_ids data is "bad", then we set it to [].  Then in
_add_untracked_ids we'd toss all the ids of things that weren't in
the tab_ids set at the end of tab_ids.  This is bad because the tabs
need to be in a specific order depending on folders.
This checkin changes _add_untracked_ids so that it builds a tab_ids
list that is correctly ordered in regards to folders.

Original sha1: 13d2b3641818865e9044ee2359f4d19698e64a2e