Bug #12197.  Return an id even after rebuilding db.
If the database is malformed or something happens to cause Miro to
throw an error when computing the last id, Miro will put the db
off to the side and start a new one in _handle_load_error.  However,
it returns None as the last id and then things fail.
This change gets Miro to try computing the last id again after
rebuilding the db.
This is a blind fix for 12197 since I'm not able to reproduce it.

Original sha1: 211043724b15d2c4d7fda8bd4a839deaffbdaf52