Bug #12059, #11987.  Fixes movies-gone code.
Overhauls portions of the startup module such that:

1. we have a better default _movies_directory_gone_handler
2. we have a better default _first_time_handler
3. the code is a little cleaner and easier to read
4. the is_movies_directory_gone function is commented
5. the check_movies_gone function creates the movies directory if
   it's not missing, but it isn't there (this happens at first time
   startup).  if that kicks up an OSError, we kick up the movies
   gone dialog.

I think the changes in step 5 will fix or alleviate the problems in
bug #12059 and #11987.

This is a blind fix because I'm not able to reproduce either of
those issues and it's not clear what circumstances lead to them.

Original sha1: f14e5c68024ff2ed85686b1fc3f83a9e6b72f851