Another fix for #12202
The issue is this:  For performance reasons, we don't load items from disk
until they are needed.  So if we migrate downloads early enough, then
downloader.itemList will be empty.  This causes items to have the wrong
filename argument and the bug pops up.

The fix is to make sure we load a downloader's items when we migrate files.

Original sha1: 950b0ff3ac70e9178066d6c5b97ca55c2d70aabf