Fixes for #12191
The first one is to make sure that we call migrate_children() when an item
still has an active downloader.  This happens often for container items, since
they are likely torrents that we are seeding.

The second is to make migrate() actually work when it's called for a child
item by calling signal_change().

Original sha1: b41504fb3fd9601de0155a5affbe205b03c62ff0