Bugs #12144 and #12275. Better table row height handling.
The initial fix for #12114 (eebfd591) occuring in 10.6 only was kind of a hack
and actually triggered a different problem on 10.4 (and maybe 10.5 too) as shown
in #12275. This change fixes both at the same time by simply providing a default
row height when none can be computed instead of noodling with the table view
initialization order.

Original sha1: 96868000d708f252b4994346a64443046ff314ff