Fixed typo when calculating the max id in upgrade88
The bug was the missing quotes and comma in this section:
'search_downloads_feed_impl icon_cache',

Replaced it with a more reliable method that doesn't involve humans typing out
a bunch of stuff.  I'm not sure if this will fix #12252, since that one
doesn't seem to have to do with icon_cache or search_downloads_feed_impl, but
it will fix similar bugs having to do with duplicate ids.

Original sha1: 8118d56e0a595a59109db52678818c0868c14963