More OS X build simplifications.
libtorrent is now the only dependency which requires boost now and it knows how
to deal with it by itself so we only have to build the required libraries in the
sandbox and be done with it. We now also build the required boost libraries
statically on 10.5 and 10.6 which simplifies even more.

Original sha1: 608d857f37cc6c40dfa18f3d988a0cf7df25862c