Fix for #12426
We should be calling signal_change() in update_status() when the state column
changes.  Calling save_now() works in most cases, except when we haven't
called save_later() and have a pending save to change, in which case it's a
no-op.

Removed redundant code setting _save_later_dc to None.  We do that in
signal_change anyway.