Fix for #12462
The issue was that if the user selected "resume seeding" before
startup_downloader() is called, then we were restarting the torrent twice.
This caused all kinds of issues stemming from using the same dlid for 2
different torrent objects.

Added checks in the RemoteDownloader class and one in the downloader daemon to
make sure this doesn't happen.