Speedup for seting up autodownload settings on a new feed.
Looping over all the items twice is definitely not what we want to do.  Added
code to support LIMITs in views and used that to find the latest item in a
feed instead of using a loop.  Made eligibleForAutoDownload=False for initial
items in a feed, since that's what's true for the all items except one.