Post binary kit update fixes on OS X.
The latest libtorrent makes our relink phase *crash*, removing some of the
command line parameters fixes this (I don't know what I hate the most, C++ or
boost). Also fixed the BOOST_ROOT variable.

In the setup.py script, exit with an explicit error when the binary kit is not
here or not up to date to prevent the (b)run.sh and build.sh scripts to continue
running.