Fixes for set_channel_title and check for updating with bulk inserts.
Realized that calling item.set_channel_title() wasn't doing much because the
item wasn't inserted in the database yet (not to mention it didn't call
signal_change()).  Fixed that by adding channel_title to the constructor
arguments.

Added a check in storedatabase that when we update an object, it actually
exists in the DB.