Speedups for creating new items during feed updates.
Removed unnecessary FeedParserValue creation.  Made Item construct input a
FeedParserValue instead of a FeedParserDict so we don't have to create them
twice.

Removed extra call to getFirstVideoEnclosure in _createItemsForParsed().
Use fp_values instead of FeedParserDict, it should be a bit faster.

Don't calculate the normalized_entry attribute in FeedParserValues.  Now that
we don't store it in the database anymore, it's just wasted work.