Bug #11923.  Fixed progresstimeline for different font sizes.
I reworked the ProgressTimeline to calculate the height it requires based
on 2 * height of a label + 14 pixels or 40--whichever is greater.  That
displays the text correctly and essentially fixes the bug.

I also fixed it to draw the background so that we don't need the display_
images anymore and the background correctly sizes to the appropriate
height.