Permformance increase when making many updates.
The change is to wrap each callback inside an SQLite transaction.  Before we
were using them only for bulk INSERT/DELETES.  Now we use them for everything,
which means that SQLite has to fsync less often, especially when we are doing
many updates at once.

The main example in my mind is moving a bunch of feeds into/out of a folder.
Before, doing this for about 8 feeds took .5 seconds on my machine.  Now I
can't see any delay.  I think there are other examples too, that was just what
popped into my head.

This also means better consistency, since if something fails we will rollback
all the operations instead of leaving us in a unknown state.

The one problem is it makes our disk full logic more complicated, since we
need to rerun all the statements in the transaction instead of just the last
one.

Also, fixed a false alarm in object_obj when the disk is full and the user
quits because of it.