Fixed a problem caused by stripper optimization.
It's possible to have text that doesn't have a < (i.e. no tags), but does have
charrefs and needs to be stripped.  I removed the optimization that screwed
up this case and added a test9.in and test9.expected testcase.

Thanks to Shot of Jaq which highlighted this glaring problem.  I apologize
profusely.