Another fix for #12680
It makes more sense to split items in the constructor than in on_db_insert().
I remember there being a reason why we couldn't do that before, but it seems
like it's fine now.  It seems going to avoid creating new items in
on_db_insert() and it's faster, since we combine everything into 1 bulk
insert.