Bug #12698.  Tweaks to moviedata_util.py to fix problems with audio files.
I tweaked moviedata_util.py so that it checks the media type before
trying to make a snapshot.  Now it only does snapshots for video
files.

I also added a second length check because the first one fails for
me all the time and I figured by then, the file is in a better
state to do a length check.  This fixes a lot of length checks for
me but it's really lame.