Bug #12328.  Wrapped kill_process in try/except.
This is a blind alleviation of the kill_process problem.  I'm not sure why
it would be kicking up an error.  My best guess is that pid is not valid
somehow or there's a race condition here where the process dies between
when kill_process is called and when it goes to kill the process.  Both
are wild guesses, though.