Bug #12699.  Adds a try-again-later possibility.
What's going on is that the file is in use when the moviedata_util goes
to check it out.  So the wait_for_play kicks up a VLC Error state and
that causes moviedata_util to fail when getting movie information
which causes the item to get tossed in the Other bin.
This fixes that.  If moviedata_util kicks up a VLC Error state,
it tells moviedata to Try again later.  moviedata sees that and doesn't
update the item details.  Miro will redo the moviedata_util run the
next time Miro is started.  In the meantime, the item falls back on
the media type it got from the first pass (checking file extension).