Bug #12711.  Fixes the select_subtitles_file plumbing for gtk platforms.
Fixed select_subtitle_file for gtk platforms so that it gets called on
VideoPlayer which calls it on the video_renderer.
Fixed the Windows code to check for the VLC Error.
Fixed the playback manager so that it checks to see if something is
playing before changing the subtitles file.