Bug #12711.  Reworks gstreamerrenderer to support external subtitle files.
This is actually a semi-big overhaul of the gstreamerrenderer.  Instead
of keeping the same playbin through Miro's runtime, we now create a new
playbin every time we select a new file.  This seems to fix a bunch of
problems with selecting external subtitles.
Also, I did some minor code cleanup while I was touching things--most of
it was pretty cosmetic.