Quicktime warm up.
On slow OS X machines, playing a movie for the first time after Miro is launched
can take a long time until Quicktime is fully initialized and prepared. To avoid
that we now asynchronously load a "dummy" movie during startup which allows to
bring Quicktime to its fully prepared state.