Fix for #12755
Two issue here, one is that we were returnning a generator when
PlaylistItemMap expected a list.  The second is that we're selecting
"MAX(playlist_id+1)", but select() assumed we passed straight column names.
Added a keyword argument to disable that code.  I think the select() code is
pretty messy right now, we should try to clean it up later.