Reduce double-subtitle actions.
Because of the way RadioActions and RadioMenuItems work, if the user selected
an item from the menu, Miro would go into convulsions disabling and enabling
subtitles.  This fixes most of that.

The menus in the menubar are still all wonky--no clue what's causing that.
The string we're putting into ui manager looks totally different from the
menu that ends up being created.