Hopefully a fix for #12538
Try to avoid calling libvlc_media_player_stop() so much when we stop a video.
Currently, we call stop() 3 times in quick succession.  When I was messing
around with this bug, I found that I could crash VLC like this.  Now we check
if VLC is already stopped and avoid sending it more messages if so.