Fix for #12509 on OS X.
Instead of deciding wether an HTTP request should be handled by Miro itself or
the browser, let WebKit start the request and let us know about the mime type as
soon as it is available. Testing the mime type then instead of the URL allows to
handle clicks on URLs which path does not contain a valid known extension.