Changed track names so they don't show up as Unknown Language.
The problem was that you could end up with more than one Unknown Language
slot in the subtitle list and there's no way to differentiate between
them and it's not obvious that the list is stable and sorted.
This changes it to differentiate between tracks and subtitle files and
includes the track id, too.