Bug #12798.  Reworked the fix so that Rename* and Remove* are in one group.
An action can be in only one action group in GTK.  So I had to rework
the menus so that menu items only have one group.
At some point in the future, we should clean up the code so that
this one group per action is enforced (i.e. fix all the "groups
is a list" related code).