Cleanup after video playback better on GTK
This change makes it so we try to destroy our persistent windows better and
unset the HWND that we send VLC after playback is finished.  This makes
improves #12692 slightly because it makes it so that VLC plays in a new
window rather than crashing.

Maybe this also fixes some other VLC hangs, I don't know.