Bug #13120.
In certain unknown circumstances, Mac OS will launch the embedded python
executable instead of the Miro executable when the application is double clicked
in the Finder, and since OS X LaunchServices pass a -psn_X_Y parameter
representing the process serial number which python does not understand, bad
things happen, as described in #13120. There is currently zero clue as to why
this happens, but the workaround is simply to move the python executable to
another location in the application bundle and update its install name
accordingly, which prevents OS X to pick it as the main executable. This also
requires to add the PyExecutableName entry to the Info.plist
(cherry picked from commit 35edd8a731abd6f4d3c61814a05d33a08cb90c59)