Bug #13061.  Changed "no results" text for video search tab.
I changed the empty-list display code so that it shows both when there
are no search results and also when there's no text in the search box.
I also added text that makes it clearer what the Video Search tab is
for and how to use it.  I think this resolves the issues in Bug #13061.