Implemented a libCURL-based httpclient.
The new libcurl subsystem runs in a separate thread from eventloop.  This
wasn't strictly necessary, but it was much simpler to implement.  The downside
is that less logic can be placed in the body_data_callback function.  A lot
the code in dl_daemon/download.py was removed and instead the smarts are in
httpclient.  For example, httpclient has support to write out to a file.

Moved ConnectionHandler and network exception classes out from httpclient and
into the new net.py module.

Updated unittest code so that we can init/destroy libcurl exactly once for the
entire unittest run.  Added local HTTP server for httpclient unittests.