Bug #13338.  Blind fix for problem with RotatingFileHandler.
If RotatingFileHandler kicks up an IOError when trying to get its
logging going, then we catch that, delete the file, and try again.

This is a blind change because I haven't figured out how to reproduce
the issue.  It might not fix the underlying issue--if
RotatingFileHandler is kicking up an IOError on a file it previously
opened, then maybe the symptoms are really indicative of bad disk?