Bug #13114.  Fixes thumbnail extraction on linux.
The problem was that there was a race condition between buffer probe and
the seek.  Therefor Miro would usually be taking a thumbnail of the first
frame which was usually black or slightly not black.

I tweaked it so that the buffer probe is added after Miro is in the PAUSED
state and after it's seeked.

I also tweaked the script so it's easier to test.