Added minimum system version support to OS X auto update.
On OS X, we delegate to (pieces of) Sparkle 1.1 to perform the auto-update
process but Sparkle 1.1 does not support the minimumSystemVersion tag which we
are going to start using with Miro 3.1 (which won't support OS X 10.4). This
patch implements custom support for the minimumSystemVersion so we don't perform
auto-update on hosts where Miro is not supported anymore.

(cherry picked from commit 80612fe991c1aa5e00a08723ad3a76455b4c2e3c)