Bug #13327.  Reduces or alleviates playback issues on Linux.
Since the problem is intermittent, it's hard to know for sure.
I was able to reproduce the problem occasionally, but not regularly
like some of the commenters in the bug.

This change made it so that I can no longer reproduce the issue
at all.  I watched and listened to podcasts in long playlists all
afternoon (and in doing so, cleared up a bunch of items sitting in
my queue) and didn't see the problem resurface.