Fix for #13390
I think what's happening is that URL is encoded poorly, non-breaking spaces
should be encoded as "%c2%a0" not "%a0".  I think what happened is that it was
encoded using latin-1, then percent encoded.  However, I believe the closest
thing to a spec is that it should be encoded as utf-8, then percent encoded.

In any case, now that we use replace, we won't crash when we see a character
that we don't understand.