Fixed selecting default subtitles on linux.
Changed code so we don't open the movie once, then open it again to play
the default subtitles.

Refactored the select_file() method a bit.  Removed the code having to do with
the subtitle file because it's only relevent for the video player.  Moved that
code to VideoRenderer.  Removed the pick_subtitle_track hack.