Bug #13256.  This fixes all the tempfile/dir shenanigans.
I fixed the tempfile/tempdirectory creation code in framework
that was already there.  Then I went through the tests and made
sure they all used the existing infrastructure.  It now cleans
up after itself.

I also tweaked the runner code so that passing in --verbose
in the args kicks it into a verbose mode which makes it sooo much
easier to see what's going on.