Fix for #13603
When find_http_auth() is called from the downloader, it must be asynchronous.
This means that I had to add back in the callback parameter, and change around
the way httpclient worked to handle the interface.

Added some cleanup code to httpclient.  In particular, this helps closing
sockets for the unittests so that there's less chance of the server hanging,
waiting for more requests.