Bug #13613.  Resumes correctly according to displayed resume time.
This also changes the resume_time saving behavior slightly.  Previously,
we wouldn't play a media item at the resume time if it was less than 10.
I moved that logic and changed it to match the mark_as_watched timeout.
So now, Miro doesn't _save_ the resume_time if it's less than 3 which
is the mark_as_watched timeout value.
Thus if Miro shows a resume time, it will resume playing at that time.