Bug #13372.  Fixes description editing.
This fixes description editing so that user-created descriptions
for feed items don't get stomped on when the feed is updated.

Additionally, it renames raw_descrption (spelled wrong) to
entry_description and changes the use-case for it.  This is now
the description from the feed.  The description column contains
the user-provided description (if there is one).  This mirrors
title and entry_title.