Fixes for #14204
First issue was displaying negative retry times.  Before we would display them
as 24 hours.  Changed it so that it said "retrying soon".  This should help
with error tracking in the future.

Second issue was that we were calling add_timeout(), then immediately
canceling the timeout.  Switched when we call cancel_request() to make the
cancel happen before we schedule the timeout.

Last issue was that we were resetting retryCount every time, so that once the
1st 2 bugs were fixed, we would still be stuck on 60 seconds.