Bug #14406.  Fixes connecting to curlmanager signals in osx.
Added some code to allow for functions to run when the curl manager
is created.  This allows the OSX frontend to connect to the
begin/end-loop signals as soon as the manager is created and allows
the manager to be created when start_thread is called rather than on
module import which is bad.