Create FileItem objects for newly-converted videos
Added ability to pass in titles/descriptions for fp_values_for_file().  Added
code in videoconversion.py that creates a FileItem using this.

Replaced the totally screwy behavior of Item._update_release_date().  The
blanket exception catching was hiding errors and it didn't make sense how we
were calling it anyway.  Now we simply update the release date for FileItems
when we set the filename.  This makes release dates work with conversions.