Tweaks to the messages sent for video conversions
This replaces VideoConversionTaskProgressed,
messages.VideoConversionTaskCompleted and friends.  It's also sent more often,
for example when an item starts running and when it fails.
This fixes a specific bug I was getting when I run a bunch of mp3 conversions.
For some reason the progress calculation is broken on that, and we were
counting on that to keep the conversions tab display correct.  I think it's
also conceptually cleaner.

Changed the VideoConversionTaskCanceled to VideoConversionTaskRemoved.  This
matches the API for other messages closer.

Menged cancel_running() and cancel_pending() into just cancel().  If a item
has started running since you've clicked the button, you probably still want
it canceled.  This matches the API for canceling downloads.

Send the VideoConversionTaskAdded message as soon as the task is added to the
pending list.