Better way to overwrite old passwords
Instead of worrying about what URLs a password will cover, just use the realm.
This is a much simpler way of handling things and it avoids certain weird
server setups where we get 2 auth headers whose domains overlap, but neither
is a subset of each other.

Made the digest auth and the basic auth's realms different in the unittests.
This is much more sane.