Bug #14624.  Changed tempfile names so they're based on filenames.
I'm pretty sure what's going on is that either the filename is getting
hosed on the command line or it's getting hosed by ffmpeg.  Whatever.
I changed it so that the filename used for the temp file (which is
what ffmpeg writes to) is based on the input file filename which
is unlikely to get hosed on the command line.