Better API for httpauth.find_http_auth
Made it synchrounous, which is much simpler to code around.

Fixed a few unittests that broke with this change.  Made the uses_httpauth
decorator start the httpclient thread in the constructor.  Before we were
calling grab_url() before httpclient.start_thread() and it was only working
because we were waiting for the httpauth callback.