Refactored config system.
Replaced module-level functions with an object that gets stuffed in app.  This
allows a more straightforward way of using a different verision in the
downloader process.

Replaced add_change_callback() with signals.  Moved the callbacks out of
config and into app.*_config_watcher classes.  This fixes threading issues
because the frontend can get callbacks on its own thread.  It also removes the
import config code in the middle of a function.

Changed how the startup code works slightly.