Optimizations for fetching ItemInfos
The idea here is to avoid hitting the object layer when fetching ItemInfo
objects.  Instead, we cache the last known value, then ask Sqlite for just the
ids of the rows we want.  This results in a huge speedup when there are many
items.

Don't duplicate infos before sending them out to the frontend.  This was
actually a fairly costly process.  Right now there's a hack to make this work
okay, I think the cleanest way would be to avoid having the fronted modify
ItemInfos.

Made LiveStorage.get_version() a public function.  The ItemInfoCache uses it
to store a variable.

Added DB code to query just ids without going through the object layer.  This
makes the caching system possible and also speeds up the Track*Count messages
substantially.