Frontend speedsup for large item lists.
Moved code out of _setup_item().  This speeds up frontend handling of large
lists of items, because we don't have to create objects for each item.
Instead, we create ImageSurface, stripped html on the fly -- caching
ImageSurfaces a bit so we aren't constantly re-creating things.

Changed 'item-added' to 'items-added'.  One signal emmision instead of many
speeds things up.

Speedup for update_menus with large lists.  We don't need to go through the
entire list to see if we can play something, 1 item is enough to know.  This
speedup only works if the list isn't entirely non-playable items, which is
fine since our largests lists are the audio/video libraries.

Optimizations for sending TableModel data to GTK