Bug #15054.  Fixes release dates for youtube entries.
YouTube feed items have both pubDate and atom:updated fields:

<pubDate>Sat, 06 Nov 2010 15:25:33 +0000</pubDate>
<atom:updated>2010-11-07T16:44:10.000Z</atom:updated>

The pub date is the original published date.  The updated date--I don't
really know what that is but my observations suggest it's always after
the pubDate both value-wise chronologically and also its placement in
the feed.

So what happens is that feedparser sees pubDate, sets updated_parsed,
then sees atom:updated and stomps on the pubDate value in updated_parsed.
Thus Miro ends up using the atom:updated date in calc_release_date
which is definitely wrong for YouTube.

I fixed this with two changes:

1. fixed feedparser to have pubDate be a synonym of published
2. tweaked singleclick.py to add published_parsed to the entry
3. HACK: tweaked item.py _calc_release_date to look at published_parsed
   for youtube feeds

The third bit is a hack, but we don't have a better system for
site-specific fixes, yet.