Fix up dodgyness in next_free_filename().
Next_free_filename() is a bit dodgy because you can race between detection
of a usable filename and the actual open of the filename.  Fix this by
doing the test and open atomically and return the filename and the
open file handle in a tuple, and let the caller deal with it.