Make writing the pid file less dodgy.
There used to be a comment to some effect like: create the pid file,
but don't truncate if it doesn't exist.  But then it goes on to write
the file from the start anyway, which doesn't make sense.

Now, make it less dodgy by removing the name from the filesystem namespace,
then create and open a new file atomically, then write the pid there.