Fix for #15312
Handle WindowsErrors when we try to move the completed file.

Added a better way to check if we are getting windows errors because the file
is in use.  Before we were just checking for errno 13, which is just
permission denied.  I can see that triggering in all kinds of situations other
than the one we want to check for.