Basic support for different frontends in OS X.
Right now I've only tested the profilewidgets frontend, but it seems to work.

Moved some code from linux startup to portable so that it can be shared with
other platforms.  Added a fix for profilewidgets frontend to make it work with
os x.  Started using optparse on OS X.

Moved code to setup theme/properties to portable This simplifies writing a new
frontend and puts all the code in one place.  I know it works for OS X and
linux, I haven't tested on windows, but we don't use startfrontend.py there
yet so that's okay.