Make qt_extractor.py a little less broken.
qt_extractor is broken, because the open command is asynchronous so
it is not allowed to do those operations until it is loaded.  Either
it will return a totally bogus value or worse yet it spins forever
while waiting for the media file to load.

Band-aid it by making it synchronous.