IPV6 fixes for windows (#15114)
Switched from a plan of always trying to create an ipv6 socket and seeing if
it worked, which was complicated and the code was buggy.  Instead, check at
once if creating a socket works or not, then use that to decide if we should
use ipv6 or ipv4.

Added a few convinence methods and cleaned the code up a bit.