Fix for #15519
The issue stemmed from changeset 5fc7b27, which works on OS X, but on
linux/windows it causes the initial navigate call not to happen.

First part is to setup an initial URL on windows so that
MiroBrowserEmbed.Enable() doesn't throw an exception.  Added a small fix for
loadURI that prevented a segfault when setting the URL to about:blank.

Second part is to change the portable code so that it doesn't depend on
viewport_created(), which is OS X only.  Added code to OS X that makes the
net-start signal be emitted in this case.