Bug #11594.  Allow people to copy item url.
This changes copy_item_url so that it can copy items that are
and aren't downloaded.

For items on the local disk, this is just the file:// url.

For items that involve flashscraping, this is the original url before it
gets converted through flashscraping.  That's fine because urls that are
converted through flashscraping are temporary by nature.