Removed obsolete code from trayicon module.
We used to do a check for gtk 2.10 or greater, but we don't need to do
that anymore since we require at least 2.16 now.  Given that, I removed
all the code that handles the case where we're running on a system that
doesn't have at least gtk 2.10.

Also did some minor PEP-8 stuff.