Bug #8655.  Overlay only shows up if hovered over.
Prior to this commit, the overlay for Miro on Linux/Windows in fullscreen
would appear whenever the mouse was moved.  Now it only appears if the
mouse is moved over the bottom of the screen.  If the mouse is moved
and not on the bottom of the screen, then the mouse will appear, but
not the overlay.