Bug #15072.  Fixes unicode/str problems with item thumbnails.
I rewrote a good portion of the thumbnail code because it sucked.  I
also added a bunch of unit tests for it mostly with the two feeds
that had problems in the bug as well as the default feeds.

It's likely there are other thumbnail situations that aren't accounted
for here.  However, this accounts for a situation that wasn't handled
previously, so Miro now gets more thumbnails than it did prior to this
commit.